Add getDisplayRegion() and data builder
This uses data from unicode-cldr/cldr-localenames-full to generate PHP
arrays containing the data.

The unit tests are currently testing every combination of locale, so
it's currently executing around 450,000 tests. The data generated is
from CLDR 27.0.1, which is the version installed on my Ubuntu 16.04
machine. Once the code has been tidied up, the unit tests will be
reduced, and the mechanism will be tested instead of the raw data.

ToDos:

 - Any @todo's marked in the code
 - Use phing to generate the metadata (similar to
   giggsey/libphonenumber-for-php)
 - Add unit tests to generate data, and run tests on that
 - Create a special unit test to run the phing tasks, and check that
   there are no differences between the committed version, and the
   generated version.